Refactored usage and definition of UserCard.js
