+++ src/components/UserCard.js
@@ -1,7 +1,7 @@
 import LinearProgress from "@material-ui/core/LinearProgress";
 import ExitToAppIcon from '@material-ui/icons/ExitToApp';
 import React from 'react';
-import {useAuth0} from "./components/react-auth0-spa";
+import {useAuth0} from "./react-auth0-spa";
 import Button from "@material-ui/core/Button";
 import Box from "@material-ui/core/Box";
 import CardHeader from "@material-ui/core/CardHeader";
